pyproject.toml
[tool.poetry]
name = "board-project"
version = "0.1.0"
description = ""
authors = ["yeeunlee <chloe.yeeunl@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
mongoengine = "^0.20.0"
Flask-Classful = "^0.14.2"
marshmallow = "^3.7.1"
Flask-Cors = "^3.0.8"
flask = "1.1.1"
Pyjwt = "^1.7.1"
bcrypt = "^3.2.0"
flask_apispec = "^0.10.0"
flask_mongoengine = "^0.9.5"

[tool.poetry.dev-dependencies]
mongomock = "^3.20.0"
pytest = "^5.2"
pytest-factoryboy = "^2.0.3"
pytest-env = "^0.6.2"
pytest-flask = "^1.0.0"
pytest-cov = "^2.10.1"
pytest-xdist = "^1.31.0"
pytest-rerunfailures = "^9.1"
flask-shell-ipython = "^0.4.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
